requirements.txt
Django>=5.2.0,<6.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
channels>=4.0.0
daphne>=4.0.0
django-cors-headers>=4.0.0
django-storages>=1.14.0
boto3>=1.26.0
Pillow>=10.0.0
python-dotenv>=1.0.0
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0
google-cloud-texttospeech>=2.16.0
mutagen>=1.47.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
whitenoise>=6.6.0
dj-database-url>=2.0.0
